crawler/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4; python_version >= '3'
colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
configparser==5.0.2; python_version >= '3.6'
crayons==0.4.0
idna==3.2; python_version >= '3'
python-dateutil==2.8.2
requests==2.26.0
selenium==3.141.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1; python_version >= '3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
webdriver-manager==3.4.2
